Implement new test structure for flexible groups
Since Flexible Groups has our focus compared to groups and it can be
considered a fully separate feature from our other group types we create
a new capabiliites folder for flexible groups and the feature of adding
content to flexible groups.

We then begin implementing our various access rules for flexible groups
and flexible group content.

+++ tests/behat/features/bootstrap/GroupTrait.php
@@ -30,7 +30,7 @@ trait GroupTrait {
     // anything in case of duplicates. Tests should prefer not having duplicate
     // titles, but testing that duplicate titles are allowed (and e.g. how path
     // aliases handles that) is a legitimate use case.
-    return end($group_ids) ?: NULL;
+    return ((int) end($group_ids)) ?: NULL;
   }
 
 }
